Add llvm text from https://llvm.org/docs/CodeOfConduct.html
I cut and pasted this into a google doc, then used an add-on tool to convert it
to markdown then tweaked that output a bit to remove internal links to that page
which will get in the way of the diffs against the FreeBSD version. I also
formatted the text so that it's clear what changes between it and the FreeBSD
version.